fix(review): split server/build.ts into focused modules (#44)
Break the monolithic buildBackroadHandler into single-concern modules under
src/lib/server: base-path (mount normalisation), index-html (per-request SPA
document), mime (download type inference), http-routes (the express Router for
auth/health/uploads/downloads/SPA) and socket-server (the lazy Socket.IO
attacher). build.ts now just wires the pieces together; types live in types.ts
and are re-exported for the long-standing import paths.
