+++ libs/backroad/src/lib/server/mime.ts
@@ -1,0 +1,24 @@
+// Minimal extension → MIME map for download_button's auto-inference. Anything
+// unknown falls back to application/octet-stream (a safe "just download it").
+const MIME_BY_EXTENSION: Record<string, string> = {
+  txt: 'text/plain',
+  csv: 'text/csv',
+  json: 'application/json',
+  html: 'text/html',
+  xml: 'application/xml',
+  md: 'text/markdown',
+  pdf: 'application/pdf',
+  png: 'image/png',
+  jpg: 'image/jpeg',
+  jpeg: 'image/jpeg',
+  gif: 'image/gif',
+  svg: 'image/svg+xml',
+  webp: 'image/webp',
+  zip: 'application/zip',
+  xlsx: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
+};
+
+export const inferMimeType = (filename: string) => {
+  const ext = filename.split('.').pop()?.toLowerCase() ?? '';
+  return MIME_BY_EXTENSION[ext] ?? 'application/octet-stream';
+};
